Regular User CRUD

+++ src/main/java/com/cooker/cook/repositories/RegularUserRepository.java
@@ -3,5 +3,5 @@ package com.cooker.cook.repositories;
 import com.cooker.cook.entities.RegularUser;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RegularUserRepository extends JpaRepository<RegularUser,Integer> {
+public interface RegularUserRepository extends JpaRepository<RegularUser,Long> {
 }
